% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability}
}
\value{
barplot
}
\description{
Title
}
\examples{
mybin(100, 10, 0.7)
}
