% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{poisson lambda}

\item{...}{additional arguments for hist()}
}
\value{
a plot
}
\description{
mycltp
}
\examples{
mycltp(n=25, iter=10000, lambda=4)
}
