% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log likelihood function}

\item{x}{data vector}

\item{param}{parameter vector}

\item{...}{extra plot arguments}
}
\value{
list
}
\description{
mymaxlik
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=10))
x=c(3,3,4,3,4,5,5,4)
param=seq(0,1,length=1000)
xlab=expression(pi)
mymaxlik(x=x,param=param,lfun=logbin,xlab=xlab,main="Binomial",cex.main=2)
}
