% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of a show from a passenger}
}
\value{
list containing provided arguments, number of tickets to sell using discrete distribution,
and number to sell using normal approximation
}
\description{
ntickets
}
\examples{
ntickets(N=400,gamma=0.02,p=0.95)
}
