% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{max x-value of shaded region}
}
\value{
curve and list of mean, standard deviation, and area
}
\description{
Title
}
\examples{
myncurve(5,1, 4)
}
